+++ src/generator.rs
@@ -386,7 +386,7 @@ impl TimeSource for StdSystemTime {
     }
 }
 
-#[cfg(any(feature = "default_rng", test))]
+#[cfg(feature = "default_rng")]
 impl Scru128Generator {
     /// Creates a generator object with the default random number generator.
     ///
@@ -499,5 +499,44 @@ mod tests_iterator {
             }
         }
         assert_eq!(i, 101);
+    }
+}
+
+#[cfg(all(test, not(feature = "default_rng")))]
+impl Scru128Generator {
+    pub(crate) fn new() -> Scru128Generator<impl RandSource, impl TimeSource> {
+        let rand_source = {
+            use rand09::{rngs::StdRng, RngCore as _, SeedableRng as _};
+
+            struct MockRandSource(StdRng);
+            impl RandSource for MockRandSource {
+                fn next_u32(&mut self) -> u32 {
+                    self.0.next_u32()
+                }
+            }
+
+            let local_var = 0u32;
+            let addr_as_seed = (&local_var as *const u32) as u64;
+            #[cfg(feature = "std")]
+            let addr_as_seed = addr_as_seed ^ StdSystemTime.unix_ts_ms();
+            MockRandSource(StdRng::seed_from_u64(addr_as_seed))
+        };
+
+        #[cfg(feature = "std")]
+        let time_source = StdSystemTime;
+
+        #[cfg(not(feature = "std"))]
+        let time_source = {
+            struct MockTimeSource(u64);
+            impl TimeSource for MockTimeSource {
+                fn unix_ts_ms(&mut self) -> u64 {
+                    self.0 += 1;
+                    self.0
+                }
+            }
+            MockTimeSource(0x0123_4567_89abu64)
+        };
+
+        Scru128Generator::with_rand_and_time_sources(rand_source, time_source)
     }
 }