+++ src/generator/default_rng.rs
@@ -1,9 +1,6 @@
 #[cfg(feature = "default_rng")]
 use rand09::{rngs::OsRng, rngs::ReseedingRng};
 
-#[cfg(all(test, not(feature = "default_rng")))]
-use rand09::{rngs::StdRng, SeedableRng as _};
-
 /// The default random number generator used by [`Scru128Generator`].
 ///
 /// Currently, `DefaultRng` uses [`ChaCha12Core`] that is initially seeded and subsequently
@@ -23,19 +20,16 @@ pub struct DefaultRng {
 
     #[cfg(feature = "default_rng")]
     inner: ReseedingRng<rand_chacha::ChaCha12Core, OsRng>,
-
-    #[cfg(all(test, not(feature = "default_rng")))]
-    inner: StdRng,
 }
 
-#[cfg(any(feature = "default_rng", test))]
+#[cfg(feature = "default_rng")]
 impl super::RandSource for DefaultRng {
     fn next_u32(&mut self) -> u32 {
         rand09::RngCore::next_u32(&mut self.inner)
     }
 }
 
-#[cfg(any(feature = "default_rng", test))]
+#[cfg(feature = "default_rng")]
 impl Default for DefaultRng {
     /// Creates an instance of the default random number generator.
     ///
@@ -49,18 +43,11 @@ impl Default for DefaultRng {
 
             #[cfg(feature = "default_rng")]
             inner: ReseedingRng::new(1024 * 64, OsRng).expect("could not initialize DefaultRng"),
-
-            #[cfg(all(test, not(feature = "default_rng")))]
-            inner: {
-                let local_var = 0u32;
-                let addr_as_seed = (&local_var as *const u32) as u64;
-                StdRng::seed_from_u64(addr_as_seed)
-            },
         }
     }
 }
 
-#[cfg(test)]
+#[cfg(all(test, feature = "default_rng"))]
 mod tests {
     use super::{super::RandSource, DefaultRng};